+++ tests/test-helpers/src/builder.rs
@@ -94,6 +94,7 @@ pub struct TestSetupBuilder {
     pub starting_block: Option<BlockHeight>,
     pub utxo_validation: bool,
     pub privileged_address: Address,
+    pub base_asset_id: AssetId,
     pub trigger: Trigger,
 }
 
@@ -187,7 +188,7 @@ impl TestSetupBuilder {
     // setup chainspec and spin up a fuel-node
     pub async fn finalize(&mut self) -> TestContext {
         let metadata =
-            SnapshotMetadata::read("../bin/fuel-core/chainspec/testnet").unwrap();
+            SnapshotMetadata::read("../bin/fuel-core/chainspec/local-testnet").unwrap();
         let mut chain_conf = ChainConfig::from_snapshot_metadata(&metadata).unwrap();
 
         if let Some(gas_limit) = self.gas_limit {
@@ -203,6 +204,9 @@ impl TestSetupBuilder {
         chain_conf
             .consensus_parameters
             .set_privileged_address(self.privileged_address);
+        chain_conf
+            .consensus_parameters
+            .set_base_asset_id(self.base_asset_id);
 
         let latest_block = self.starting_block.map(|starting_block| LastBlockConfig {
             block_height: starting_block,
@@ -247,6 +251,7 @@ impl Default for TestSetupBuilder {
             starting_block: None,
             utxo_validation: true,
             privileged_address: Default::default(),
+            base_asset_id: AssetId::BASE,
             trigger: Trigger::Instant,
         }
     }